Refactor codebase into better packages
