+++ src/main/java/poseidon/Main.java
@@ -8,6 +8,7 @@ import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import poseidon.ui.MainWindow;
 
 /**
  * A GUI for Poseidon using FXML.
