+++ src/main/java/poseidon/storage/Storage.java
@@ -1,4 +1,4 @@
-package poseidon;
+package poseidon.storage;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -10,6 +10,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Scanner;
 
+import poseidon.exception.PoseidonException;
 import poseidon.task.Deadline;
 import poseidon.task.Event;
 import poseidon.task.Task;
