+++ src/main/java/poseidon/tasklist/TaskList.java
@@ -1,4 +1,4 @@
-package poseidon;
+package poseidon.tasklist;
 
 import java.util.ArrayList;
