+++ src/main/java/poseidon/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package poseidon;
+package poseidon.ui;
 
 import java.io.IOException;
